+++ src/core/mod.rs
@@ -1,1 +1,6 @@
 pub mod buffer;
+pub mod program;
+pub mod rendertarget;
+mod shader;
+pub mod state;
+pub mod texture;
