+++ src/lib.rs
@@ -7,17 +7,12 @@ pub use gust::mesh;
 pub mod core;
 pub mod loader;
 pub mod model;
-mod shader;
 mod utility;
 
 pub mod light;
 pub mod material;
-pub mod program;
 pub mod input;
-pub mod state;
 
-pub mod rendertarget;
-pub mod texture;
 pub mod eventhandler;
 pub mod camera;
 pub mod scene;
